compleated level_12.lesson_4.Task_2

+++ src/main/java/level_12/lesson_4/Task_2.java
@@ -3,7 +3,7 @@ package level_12.lesson_4;
 public class Task_2 {
     public static void main(String[] args) {
         print(1);
-        print(12);
+        print((Integer) 12);
     }
 
     public static void print(int i) {
